+++ src/basic/ui/index.ts
@@ -1,3 +1,4 @@
 export * from './notification-item';
 export * from './search-input';
 export * from './admin-dashboard';
+export * from './product-list';
